[bug] fixed bug with env vars from exec-env
they were not being set if entering a directory that also had a
.tool-versions file inside of it that exported the same environment
variable.

e.g.: if 2 .tool-versions files both had golang in it, entering the
second directory would cause GOROOT to be _cleared entirely_.

The bug was that I was using the _current_ environment here when I
should have been using the _pristine_ environment.