bug: handle .ruby-version and Gemfile properly
I had not considered what would happen when there were 2 legacy versionfiles, which is likely with ruby when you would have `.ruby-version` and `Gemfile` in the
same directory.

This fixes 2 bugs with that situation, first, if the Gemfile did not
specify a ruby version the script returns nothing. In that case, the
legacy file should fall through and it should go to .ruby-version or
.tool-versions ignoring the Gemfile.

The other is that I was using HashMap for the legacy version files
which ignored the ordering specified by the plugin. `.ruby-version`
is supposed to be parsed _before_ `Gemfile`. In the previous behavior
that was nondeterministic