feat: bake shortname repository into codebase
The shortname repository can be persisted into a simple file. This
is not a significant amount of data so the cost is minimal.
I suppose the downside is this means any changes will need to wait
for a new rtx release, but I do not see that as a major problem.

As a bonus, I was able to add an alias for node so now you can use
node or nodejs and rtx will work the same way