feat: display current tool-versions value if no version set (#132)
Fixes #67

Before this change the behavior of something like `rtx local ruby` was
pretty inconsistent. Generally it would set the version to what was in
the global tool-versions file or just error out. Either one seemed
pretty useless to me.

I was thinking this could be a "get" instead of "set" mode, so now it
will display what the current value is in `.tool-versions`. This is
slightly different than `rtx current ruby` because that shows what is
the currently used which could be coming from any .tool-versions file
(or env var or something). This _only_ shows the value from the current
`.tool-versions` file.

I'm not entirely sure this behavior is good/predictible, so I may regret
this decision but I think it is at least better than what we have
currently.