Avoid panic if run_with_timeout hits timeout.
If run_with_timeout hits the timeout, the mpsc receiver may be
dropped, leading to a panic when the background thread later
attempts to return its value.

This change simply ignores any error. Since the timeout has
already been reached, reporting on this error would just be
noise.

A future improvement could be to cancel the background operation,
but that would require a much larger refactoring.

Part of #860