[bug] Fix `mise x cargo:eza` not installing during run
Addresses a bug with `mise exec` where using an unversioned tool
reference causes the underlying binary to not be installed on first use.

Instead we treat an unqualified tool which has no current install as having tool@latest
and handle that in the else clause.

The desired behavior is:
eg mise x cargo:eza
- If any version of cargo:eza exists, use it
- If no version exists, install latest