Fix `mise shell` setting env in nushell
Currently `mise shell` doesn't have an effect in latest nushell.
Functions in nushell that want to export environment variables
must use the `--env` flag in the function definition.

This flag was used on other functions in mise.nu as required
(eg. update-env), however it seems that this flag is also required
on the `main` function that manages updating the env for `mise shell`.