fix: ensure `mise install` installs missing runtimes listed in `mise ls`
Previously `mise ls` could show missing runtimes, but `mise install`
would report all runtimes were installed. The latter only checked if
the install directory existed, which could be present if `mise
install` aborted early. However, the former also checked the existence
of the runtime symlink and the `incomplete` file in the cache.

To make this behavior consistent, dispatch the tool request to the
backend. Exclude the runtime symlink check since this breaks the
install check for `node@latest`.

Closes https://github.com/jdx/mise/issues/2257