fix: Fixes fish_command_not_found glob error (#4133)
An error for `fish_command_not_found` was introduce in #4094 because
glob misses do not work the same in fish as in other shells. As such,
we are getting the following error on an unknown function:

```
- (line 64): No matches for wildcard '"mise-"*'. See `help wildcards-globbing`.
    if test "$argv[1]" != "mise" -a "$argv[1]" != "mise-"*
```

This has been replaced with a fish-specific implementation test that
prevents the error and is somewhat more idiomatic as well.