fix: mise does not operate well under Git Bash on Windows (#4048)
Fixes #4011 

Gave this a try but detection of `msys` or `cygwin` does not work via
`$OSTYPE` as this is not available via `env` so `env::var()` does not
return anything. Not sure how reliable my approach via `MSYSTEM` and
`PWD` is though.

Another unresolved issue seems to be related to line breaks on cygwin.
The `.bashrc` file does not source successfully until its opened and
saved (without changes). Probably would require a `dos2unix` on cygwin,
not sure why that is not an issue on msys.
```bash
$ source .bashrc
': not a valid identifiernctions
-bash: .bashrc: line 35: syntax error near unexpected token `$'\r''
'bash: .bashrc: line 35: `function __zsh_like_cd() {
```