fix: lowercase desired shim names (#4333)
This fixes a bug on macOS where we are searching for a shim name with
one or more characters uppercased, e.g. Python. But on macOS the APFS
(default file system) is case insensitive, so `mise doctor` will always
report an error in this case.

See https://github.com/jdx/mise/discussions/4332