feat(registry): add 'marksman' via 'aqua:artempyanykh/marksman' backend (#4357)
This PR adds [marksman](https://github.com/artempyanykh/marksman) to the
Mise registry via the Aqua backend (see PR
[#31906](https://github.com/aquaproj/aqua-registry/pull/31906), merged
as of yesterday).

Marksman implements the Language Server Protocol for Markdown to be used
in terminal- and non-terminal-based text editors. Marksman is developed
in F# and distributed as a single binary for Linux, MacOS, and Windows.
Installation using `mise` was tested successfully on Fedora and Ubuntu.

---------

Co-authored-by: jdx <216188+jdx@users.noreply.github.com>