refactor: migrate humantime to jiff (#4616)
Follow up of #4612. `humantime` crate is now outdated, so migrate to
[jiff](https://crates.io/crates/jiff) instead as `cargo-deny` suggests.

`jiff` is almost compatible with `humantime`, but is not 100% compatible
as documented.

https://docs.rs/jiff/latest/jiff/fmt/friendly/index.html#comparison-with-the-humantime-crate
Notably, `1M` is parsed as `1 month` by `humantime` but cannot be parsed
by `jiff`.

I used `2025-01-01` when converting `Span` to `SignedDuration` since we
need a relative date to convert `months` or `years`, in which the
duration differs depending on dates.

https://docs.rs/jiff/latest/jiff/struct.Span.html#integration-with-stdtimeduration-and-signedduration

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>