feat(aqua): support github_release minisign type (#4897)
Fixes https://github.com/jdx/mise/discussions/4895.

ref:
https://github.com/aquaproj/aqua/blob/a3e7356995673ed372b69d20dda148cd4b963171/pkg/config/registry/minisign.go#L3-L11

The following three packages use `minisign`.
However, `bufbuild/buf` signs a `sha256.txt` file, which is not
supported, so I only checked for the other two:
- `ziglang/zig`
- `zigtools/zls`
- `bufbuild/buf`  

https://github.com/search?q=repo%3Aaquaproj%2Faqua-registry+minisign%3A&type=code

---

For SLSA, `type` is parsed as `String`, but for checksums it's parsed as
the `AquaChecksumType` enum.
I added `AquaMinisignType` enum for consistency, but if using `String`
is preferred, I can update it.

https://github.com/jdx/mise/blob/46a58953725853388aebfe4060e3ed18a80c32e4/src/aqua/aqua_registry.rs#L131-L133

https://github.com/jdx/mise/blob/46a58953725853388aebfe4060e3ed18a80c32e4/src/aqua/aqua_registry.rs#L151-L152

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>