perf: turn several of the list functions into parallel iters (#4924)
`mise up` is quite slow if you have a lot of tools (especially go tools)
so parallelize the outdated checks. While I was at it I turned a few of
the other `iters` into `par_iters` but I can switch them back if you
want (the outdated one was the main issue).