feat: add check flag for the fmt command (#4972)
Add a `--check` flag for the `fmt` command. When set no formatting is
done.

While adding this i wondered what the following should do. This does
most likely not work as intended as the `sort_values_by` clearly states
that it does NOT operate on header keys as the ordering in the order
closure suggests.


https://github.com/jdx/mise/blob/576014ff938e4ca77f5e07a0c9d3f751cf4d1678/src/cli/fmt.rs#L65-L84

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>