docs: fix add_predicate handler in neovim cookbook (#5044)
According to [the
documentation](https://neovim.io/doc/user/treesitter.html#vim.treesitter.query.add_predicate()),
the TS predicate handler function is supposed to return boolean (or
nil). It previously returned either a string or nil depending on whether
the filename matched the predicate or not. This caused the example to
not work (at least on my machine running nvim 0.11).

Fixed it by adding a nil check for the `string.match` return value.