feat: use new Java metadata source (#5089)
This is draft to use a new Java version data source. Its output is
currently tailored to `mise` but could provide more extensive data
(architectures, properties etc.) if need be.

It is using [roele/roast](https://github.com/roele/roast), a rewrite of
[jdx/rtx-java-metadata](https://github.com/jdx/rtx-java-metadata)
respectively
[joschi/java-metadata](https://github.com/joschi/java-metadata).

Improvements:
* contains only `mise` relevant data/properties
* more consistent versioning
* uses proper vendor provided checksums

Possible breaking changes:
* removed AdoptOpenJDK (has been replaced by Temurin in 2021)
* few patch versions are missing as they are not provided by the vendors
anymore


Question is if we should fork/move the repository for `mise` purpose to
somewhere else and if we should push the data to a domain that is more
related to mise. @jdx WDYT?

/cc @bric3