feat: add shims in REMOTE ENV (#5139)
Some VS Code extensions (e.g., Terraform) do not function properly
within the DevContainer environment. Specifically, they do not provide
an option to configure the path to the tool's binary, which leads to
issues when the tool is installed in non-standard locations (e.g., via a
version manager like mise).