fix: rayon -> tokio (#5172)
The deadlock issue reported by many people but the cause seems related
to using rayon for parallelism. This PR switches to using async rust
with tokio which should resolve it.

This is obviously a massive PR so don't be surprised if there is a bug
in some untested code path.

Fixes #5149
Fixes #5144 
Fixes #4302
Fixes #4127