fix: installing tools with postinstall hooks fails (#5191)
According to the discussion #5001 the change #4960 to support GH
Enterprise/GitLab is causing a regression for certain tools (e.g. aqua)
because of the global ToolOption argument `postinstall`. To address this
we should filter the global once on install arguments and only store
ToolOptions for certain backends (`cargo`, `go`, `pipx`, `ubi`)