fix: set handler for ctrlc on windows shell (#5209)
## Background

Windows way of creating a process is to just go ahead and pop a new
process with given program and args into existence. But in unix-land, it
instead happens in a two-step process where you first fork the process
and then exec the new program, essentially replacing the current process
with the new one.

Now, when using `mise exec` in windows, it spawns a child process,
instead of being able to replace itself like in unix impl using the
`.exec()`, which makes a `exec` syscall.

So visually, it looks a bit like this:

```
Parent Process
  |
  |— Child Process (say erl or iex)
```
Now when we send a `Ctrl + C` input to console, it first hits parent
process, which then makes parent quit. Then it propagates to child,
which can chose to handle `Ctrl + C` by not exiting, such as in case of
`erl`, where a break menu is shown instead.

Here comes the kicker, in Windows, orphaned processes are not cleaned
automatically, which causes it to be left behind, and remain there as a
ghost process. And if this ghost process is able to do interactive IO on
console, we are in for a session of ghostbusters.


![image](https://github.com/user-attachments/assets/fdd57ccd-2bad-4fae-84ef-ddbc6fa71c4e)

Now for the particular case of `erl` based elixir shell, it ends up
messing with console since the freshly orphaned process now acts in
weird ways, being able to handle and interfere with the current console,
creating a race condition where the host process for the console, and
this parentless process both compete to do stuff in console.

Refer #5174

## Proposed Fix

Just make it so child process is never parentless. When doing exec,
ignore Ctrl + C to parent, and let the child deal with it. If the child
dies, parent follows and exits.

To achieve this, we use
[SetConsoleCtrlHandler](https://learn.microsoft.com/en-us/windows/console/setconsolectrlhandler)
to ignore `Ctrl + C` on parent. We don't outright disable it by sending
a `(NULL, TRUE)` to it since it otherwise is not propagated to the
child. This lets `iex` work as expected on `Ctrl + C`.


![image](https://github.com/user-attachments/assets/7870eabd-e930-4f47-9f97-564de7f92ba1)

Here, `./mise.exe` is a binary I built with my version of the fix, which
is now working as expected. Only caveats of this fix would be needing to
dig into `unsafe`, and new dependency on `winapi`, which ends up boiling
down to preference.

(Now here, there's this prompt for `Terminate batch job (Y/N)?` which
comes because `iex` is internally a bat file on windows, and works as if
it was installed directly. It is expected.)