fix: Task resolve_depends for multiple depends_post values
resolve_depends was iterating over all tasks for traversal when processing depends_post.
When a task had a single post_depends, this was mostly fine. However the issue was most visible
with multiple post_depends values, where an infinite loop in the graph was generated, however
the actual dependencies were not generated leaving the tasks listed in post_depends un-executed.

This lead to the following error:

[a] panic in task: panicked at src/task/deps.rs:101:13:
Infinitive loop detected, all tasks are finished but the graph isn't empty

By initially iterating over self.depends_post we properly resolve the dependencies