fix(aqua): parse consecutive pipes in aqua templates (#5385)
Resolves https://github.com/jdx/mise/discussions/5382.

If the templates have multiple pipes, only the first and last functions
are parsed correctly.
```
trimV .Version | replace "-RC" "-rc-" | replace "-M" "-milestone-"
```
This is parsed as `9.0.0-RC1` if `.Version` is `9.0.0-RC1`, but should
be parsed as `9.0.0-rc-1`.

This PR fixes the issue by adding `in_pipe` state in the `parse`
function.