Use os_release data, scope to Ubuntu versions
Use the os_release crate, which is already loaded in (and used in Swift)
to determine what OS we're on.

BOB only supports ubuntu 20.04, 22.04, and 24.04, so we return an
Ok(None) if we're not on one of these versions. The Ubuntu check is
relatively modular, so should Bob add more support later, or other
mirrors pop up (fedora, etc), we can add those easily