fix(erlang): use precompiled binaries for linux ubuntu (#5402)
This PR attempts to enable installation of precompiled erlang binaries
on Ubuntu systems, supplementing the current fast installs for MacOS and
Windows (#2721).

# Details

Erlang builds are made available, precompiled, for a variety of Ubuntu
systems via [bob](https://github.com/hexpm/bob).

These are generally available at
`https://builds.hex.pm/builds/otp/${ARCH}/${OS_VER}/${REF}.tar.gz`

`ARCH` and `OS_VER` are typically a bit different than the other
constants Mise uses in the erlang.rs core plugin.

The `REF` appears to be the same as the ones used for the other
functions in the file, so it does not need to be changed.

From bob's documentation, the two arches available for precompiled
builds are:

- `amd64`
- `arm64`

Additionally, the OS version encodes what build of Ubuntu is used as
well:

- `ubuntu-20.04`
- `ubuntu-22.04`
- `ubuntu-24.04`

The architecture names are a fairly simple mapping, more or less just
different names for concepts that already exist in the mise source code.
However OS_VER is more complicated.

On Github Actions, this is exposed as the envar `ImageOS`. Determining
the runtime from that would technically satisfy my ask, as this is
primarily about CI.

In this PR, I check the value of the `ImageOS` var, and if its absent
use the `os_release` crate, which is already loaded for use elsewhere,
to fetch the version. Since, as mentioned earlier, bob only has builds
for a few ubuntu versions, I check against those, and if either ImageOS
or the os_release data don't match, return early so that we can still
build manually

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>