feat(registry): use ubi backend for youtube-dl nightly releases (#5466)
> This release is old and has many known issues. See "... commits to
master since this release" above.
> You almost certainly want the [nightly build
release](https://github.com/ytdl-org/ytdl-nightly/releases/), and not
this one. The nightly build is created from the master branch of this
repo if changes have been committed since the last build. Read about
[Installation](https://github.com/ytdl-org/ytdl-nightly#user-content-installation)
of the nightly build. Major fixes in the nightly build are listed
https://github.com/ytdl-org/youtube-dl/issues/30839.
ref: https://github.com/ytdl-org/youtube-dl/releases/tag/2021.12.17

According to the latest release, we should use `ytdl-org/ytdl-nightly`.
However, the test fails since the versions host don't include the
nightly build versions, I disable it for now.

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>
Co-authored-by: jdx <216188+jdx@users.noreply.github.com>