feat(registry): import package descriptions from aqua and add os specifier for tuist (#5487)
Noticed that only 3 packages in the registry had descriptions, so took
the opportunity to pull descriptions across for all packages that have
an aqua backend to help with discovery.

Also made an update to `tuist` to have an os specifier for macos since I
was getting errors on linux/windows when it couldn't be installed.

-
[`401e8a39`](https://github.com/jdx/mise/pull/5487/commits/401e8a39cd9fb8834377b33efb437a799921a76b)
chore(registry): import description from aqua for package with aqua
backend
-
[`834c0e2e`](https://github.com/jdx/mise/pull/5487/commits/834c0e2e582741eb33c2288fa7d95d6682931fe6)
fix(registry): add os specifier and description for tuist