fix(aqua): apply correct `version_override` (#5474)
Fixes the issue I wrote in
https://github.com/jdx/mise/pull/5371#issuecomment-2991726382.

The cause of the issue was that mise tried to find the
`version_override` with a v-prefixed version.
For example, even if the input was the full GitHub tag like
`@biomejs/biome@2.0.0-beta.5`, it tried to find the override for
`v@biomejs/biome@2.0.0-beta.5`.
This led to it finding the fallback override, which has `cli/` for the
`version_prefix`.
The simple fix is to find the override without the v-prefix.

However, this would break packages that rely on the v-prefix in the
`version_constraint`.
So, I ended up finding the override that matches either the input
version or the v-prefixed version.
This means it tries to find the first override that matches one of
`v@biomejs/biome@2.0.0-beta.5` or `@biomejs/biome@2.0.0-beta.5` for the
previous example.
In this way, we can correctly find the override for most packages.

Also, this PR removes the retry logic when a GitHub tag is found from
`get_version_tags` or if the input version already has a `v` prefix.
This will allow it to fail fast and also simplify the error log.

Package | Version String | `2025.6.8` | #5474
-- | -- | -- | --
biome | 2.0.0 |✅ | ✅
biome |
[@biomejs/biome@2.0.0](https://github.com/biomejs/biome/releases/tag/%40biomejs%2Fbiome%402.0.0)
(tag) | ❌ | ✅
biome | v2.0.0 | ❌ | ❌
biome | 2.0.0-beta.5 | ❌ | ❌
biome |
[@biomejs/biome@2.0.0-beta.5](https://github.com/biomejs/biome/releases/tag/%40biomejs%2Fbiome%402.0.0-beta.5)
(tag) | ❌ |✅
biome | v2.0.0-beta.5 | ❌ |❌
biome | 1.9.4 | ✅ | ✅
biome | v1.9.4 (tag without prefix) | ✅ | ✅
biome |
[cli/v1.9.4](https://github.com/biomejs/biome/releases/tag/cli%2Fv1.9.4)
(tag) | ❌ | ✅
delta |
[0.18.2](https://github.com/dandavison/delta/releases/tag/0.18.2) (tag)
| ✅ | ✅
delta | v0.18.2 | ❌ |❌
eza | 0.21.4 | ✅ | ✅
eza |
[v0.21.4](https://github.com/eza-community/eza/releases/tag/v0.21.4)
(tag) | ✅ | ✅

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>