fix: `self-update` on MITM firewall (attempt #2) (#5459)
This is a follow-up to jdx/mise#5387, to fix the `self-update` command
when using a traffic-inspecting/MITM firewall, such as Cloudflare WARP.

The previous PR broke the release build, so had to be reverted. I didn't
realise that the [`reqwest`](https://crates.io/crates/reqwest) crate
requirements are different on Linux (I use macOS), so OpenSSL is still
needed.

I tried removing OpenSSL as a dependency completely, but it appears to
not be straightforward, as many other crates also depend on `reqwest`.

The simplest thing I think is just to leave OpenSSL and add the
`rustls-native-roots` feature.

I've tested `build-tarball.sh` locally on both macOS and Linux (using a
VM) and verified that the resulting binaries can self-update correctly.

I don't have an easy way to test on Windows, but the `reqwest` docs
suggest that it behaves the same as on macOS.