fix: do not overwrite github tokens environment variables (#5546)
Fixes https://github.com/jdx/mise/issues/5524.

https://github.com/jdx/mise/pull/3114 introduced the overwrite behaviour
to resolve https://github.com/jdx/mise/issues/2910 and
https://github.com/jdx/mise/issues/3070.
It was to ignore an invalid `GITHUB_TOKEN` if `MISE_GITHUB_TOKEN` is set
to an empty string, but it was necessary only because we used to call
`ubi` from the CLI.

For asdf plugins, this should be enough to overwrite `GITHUB_TOKEN`.

https://github.com/jdx/mise/blob/ce78f04bfb338d218df90301a093b01b1d67d13a/src/plugins/asdf_plugin.rs#L426

I'm not sure about vfox plugins, but I don't think this PR will change
any behaviours because it seems `vfox.rs` not using the tokens.