fix(aqua): align version resolution logic in list_bin_paths (#5562)
Aligns the version resolution logic (from canonicalised version to
GitHub tag) in `list_bin_paths` with the one in `install_version_`.

I think this doesn't change the behaviour for offline introduced in
https://github.com/jdx/mise/pull/5195.

> TODO: instead of caching it would probably be better to create this as
part of installation

Regarding this TODO, I agreed with it and tried to work on it, but was
not sure how to.
I'd like to ask for some advice.

I thought of storing `bin_paths.msgpack.z` in the installation path,
`tv.install_path()`, while the installation, and reading it when
`list_bin_paths` is called.
Not to break the installations with old versions, or just in case
`bin_paths.msgpack.z` is removed somehow, I think we need to fall back
to the current implementation if it's not found.

However, then it would be the same as the current implementation, except
that `list_bin_paths` is called in the installation and the cache is
created.