fix(cli): stable path env for exec on windows (#5790)
This PR resolves #5588. It fixes the issue on Windows, where `mise x`
randomly provides a different `PATH` environment variable to the tool it
executes.

@Wazzaps demonstrated that this issue is caused by some undefined
behavior in the [duct.rs](https://github.com/oconnor663/duct.rs)
library. The following PRs might solve the issue:

1. https://github.com/oconnor663/duct.rs/pull/119
2. https://github.com/oconnor663/duct.rs/pull/120

This PR offers an alternative solution to the issue, by simply avoiding
the problematic `duct.rs` API, that is `cmd.env(k, v)`, and setting the
environment ourselves, i.e. `env::set_var(k, v)`, exactly how it is done
for the Unix implementation of `exec`. When `duct.rs` launches the child
process, the tool inherits the environment as-is.

Since `exec` on Windows implements a close approximation of the Unix
`exec` behavior, setting the process environment of `mise` should be
fine, since `mise x` will only run for the duration of the child tool
process.

An e2e-win test is added to demonstrate that the `PATH` environment
variable is now stable between successive runs of `mise x`.

Previous PRs related to `mise x` behaviour on Windows:

1. #2648 
3. #2598 
4. #5209

---------

Co-authored-by: Willie Möller <willie.moller@5dt.com>