fix: misidentifying built-in backend as a plugin backend  (#5822)
Fixes: https://github.com/jdx/mise/discussions/5646

Issue introduced here:


https://github.com/jdx/mise/compare/v2025.7.7..v2025.7.8#diff-666bb9d34c6ca6727005c50be93f02d720e77f525dcef4d8d286344c1f0d21f1R101-R126

The new backend logic was causing the `versions_hash` in
`get_leaf_dependencies` to always include "pipx" when testing with a
pipx tool. This causes issues when pipx is not part of the requested
tools to install. It essentialy prevented the tool from being considered
a leaf dependency.