docs: fix uv_venv_create_args reference for python (#5854)
Hey, disclaimer is I am a noob with `uv`, but when setting uv up
following the mise + uv configurations, I kept having errors with
`uv_create_args`, it appears to me that it is a typo and should actually
be `uv_venv_create_args`.

Hopefully this is helpful for others. Feel free to merge or close if
this is incorrect, no worries.

Thanks for this awesome tool. Been using since the rtx days.