registry: use aqua backend for yarn (#5902)
Sorry, I didn't notice the author hadn't updated the PR to use the aqua
backend in https://github.com/jdx/mise/pull/5745.

I didn't notice that the registry supports `idiomatic_files`, but it
does, so `.yvmrc` will still be supported.
`.yarnrc` was in the docs, but it had never been supported by the asdf
plugin, so I guess it was a mistake.

yarn v1 is no longer supported since
https://github.com/jdx/mise/pull/5745, but we can still install it with
`asdf:yarn` (not on Windows).

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>
Co-authored-by: jdx <216188+jdx@users.noreply.github.com>