fix: downloading musl node binary would fail and try to compile even when exists on custom mirror
The main issue was that my settings.node.compile false was being overridden by the distro all_compile override, but to get to the bottom of it I had a few bugs and weird behavior, so I restructured the installing portion of node while at it. The fix was avoiding the override if node.compile was set by the user manually in config/settings.rs.

The rest of the changes in plugins/core/node.rs, I made a fetch_binary to reuse code for posix and windows, made sure the errors where not encapsulated by using Report::from (I had a bug because of it while testing stuff that shouldn't happen), made the logic about Ok explicit and about the NotFound case too. Also made the unarchiving logic reusable, I was thinking in moving to another place to let other packages use something like this, but it's just that for now.