docs: Update install instructions for nushell (#5981)
[This PR](https://github.com/jdx/mise/pull/3592) should have fixed the
path issues, but the problem described in [this
discussion](https://github.com/jdx/mise/discussions/3555) was not fixed.

The issue doesn't persist if the file is autoloaded, but the
documentation currently describes the old way of importing the file with
use. This updates the docs to write the autoload-way into the nu config