registry: rename benthos to redpanda-connect (#5984)
As pointed out in https://github.com/jdx/mise/discussions/5980,
[`benthosdev/benthos`](https://github.com/benthosdev/benthos) now
redirects to
[`redpanda-data/connect`](https://github.com/redpanda-data/connect).
They renamed it due to an acquisition, as in
https://www.redpanda.com/press/redpanda-acquires-benthos.
https://github.com/redpanda-data/connect/pull/2618

`aqua-registry` already supports it, so we can safely rename it.
https://github.com/aquaproj/aqua-registry/pull/23450
[`benthosdev/benthos-asdf`](https://github.com/benthosdev/benthos-asdf)
supports the latest version of `redpanda-connect`, so I think we can
keep this. As a note, it renames the installed bin to `benthos`.

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>