registry: rename coq to rocq (#5985)
As in https://github.com/jdx/mise/discussions/5980,
https://github.com/coq/coq has been renamed to
https://github.com/rocq-prover/rocq.

The asdf plugin uses `coq/coq`, but it will be redirected, so it should
be fine. I will open a PR to rename it too.

https://github.com/mise-plugins/mise-coq/blob/873cda47525cb72b88e516108d0b8c6b1a5e4e96/bin/install#L93

I also added the dependencies, but they also need `opam install -y num
ocamlfind camlp5`, so it wouldn't work.

https://github.com/mise-plugins/mise-coq/blob/master/README.md#prerequisites

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>