fix(nushell/windows): Fix $env.PATH getting converted to a string (#6157)
This addresses the issue in discussion #5472.

Mise on nushell on windows doesn't properly convert the path to a
nushell-native list. Most of the time this works fine, as long as the
path stays a string. However, if anything else adds a value to the path
(like manually activating a python venv), you end up with a list of 2
items: the big, `;` delimited string that was $env.PATH, and one entry
for whatever was just added (now the only valid item in the path list).

The problem comes from this function in the activate script:

```nu
def --env "update-env" [] {
     # ...
     if $var.name == 'PATH' {
        $env.PATH = ($var.value | split row (char esep))
     } else {
    # ...
}
```

`$env.PATH` is case-insensitive in nushell [1] (actually, all
environment variables are, regardless of OS [2]). Mise matches whatever
capitalization is in use at run time [3], which for windows is `Path`.
For the sake of the update hook though, in case something changes the
capitalization of `$env.PATH` between activating mise and some later run
of the hook, it's probably just better to do this comparison in a
case-insensitive way.

[1]: https://www.nushell.sh/book/configuration.html#path-configuration
[2]: https://www.nushell.sh/book/environment.html#case-sensitivity
[3]:
https://github.com/jdx/mise/blob/b7583ea9cc43c78e60d044640d149958e71261e4/src/env.rs#L225C1-L231C4

Co-authored-by: Zack Yancey <mail@zackyancey.com>