fix: use config_root for env._.path (#6204)
The behavior and docs for env._.path predates config_root as well as
config files outside of the root like `.config/mise.toml`. This has been
a bug and should have been fixed when support was added for files
outside the project root but was not. This change modifies the behavior
of env._.path to behave the same as other parts of mise that are
relative to config_root and not the parent directory of the config file
making mise behave consistently in regards to relative paths.

Fixes https://github.com/jdx/mise/discussions/6203

BREAKING CHANGE: env._.path directories are now relative to the
config_root and not the parent directory of the config file. This only
affects config files that use env._.path that are outside the
config_root (e.g.: `.config/mise.toml` will use `.` and not `.config`
like before). To retain compatibility with earlier mise releases, be
explicit with `env._.path = "{{config_root}}/bin"` which will behave the
same in old and new releases.

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>